composable_app/pyproject.toml
[project]
name = "composable-app"
version = "0.1.0"
description = "Multi-agent content writing system demonstrating enterprise GenAI patterns"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai",
    "streamlit",
    "mem0ai",
    "chromadb",
    "sentence-transformers",
    "python-json-logger",
    "nest-asyncio",
    "llama-index-llms-openai",
    "llama-index-embeddings-openai",
    "llama-index",
    "openparse",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "utils", "pages", "prompts", "data", "evals"]

